fix: controller

+++ src/controllers/lanches.controller.js
@@ -7,12 +7,26 @@ const findLanchesController = (req, res) => {
 
 const findLancheByIdController = (req, res) => {
   const idParam = Number(req.params.id);
+  if (!idParam) {
+    return res.status(404).send({ message: "Lanche não encontrada!" })
+  }
   const chosenLanche = lanchesService.findLancheByIdService(idParam);
   res.send(chosenLanche);
 };
 
 const createLancheController = (req, res) => {
   const lanche = req.body;
+  if (
+    !lanche ||
+    !lanche.local ||
+    !lanche.localizacao ||
+    !lanche.nome ||
+    !lanche.descricao ||
+    !lanche.preco ||
+    !lanche.foto
+  ) {
+    return res.send({ mensagem: "Você não preencheu todos os dados para adicionar uma nova paleta ao cardápio!" });
+  }
   const newLanche = lanchesService.createLancheService(lanche);
   res.send(newLanche);
 };
@@ -20,12 +34,22 @@ const createLancheController = (req, res) => {
 const updateLancheController = (req, res) => {
   const idParam = +req.params.id;
   const lancheEdit = req.body;
+  if (!idParam) {
+    return res.status(404).send({ message: "Lanche não encontrado!" })
+  }
+
+  if (!lancheEdit || !lancheEdit.local || !lancheEdit.localizacao || !lancheEdit.nome || !lancheEdit.descricao || !lancheEdit.foto || !lancheEdit.preco) {
+    return res.status(400).send({ message: "Você não preencheu todos os dados para editar a paleta!" });
+  }
   const updatedLanche = lanchesService.updateLancheService(idParam, lancheEdit);
   res.send(updatedLanche);
 };
 
 const deleteLancheController = (req, res) => {
   const idParam = req.params.id;
+  if (!idParam) {
+    return res.status(404).send({ message: "Lanche não encontrado!" })
+  }
   lanchesService.deleteLancheService(idParam);
   res.send({ message: 'Lanche deletado com sucesso!' });
 };
